Temporarily resolved conflict from missing mount
That is the GPS/IMU mount.  This resolution leaves the GPS and IMU parts
floating in the swarmie assembly.  Will need to mount new GPS/IMU Mount
asssembly to the top plate, then reattach the GPS and IMU parts to the
mount.

This will be finally resolved once the GPS/IMU mount is completed to
satisfaction.