Finalized cow catchers
There are now two versions.

One is intended for the rear of the robot, to
prevent the robot from climbing onto objects when reversing or turning.

The next is to prevent the robot from climbing onto objects in the front.
This version is designed specifically to prevent the gripper from climbing onto blocks.